Implement group close by prio fee rate
It makes no sense to maintain groups that pay higher commissions than those required to enter the next block. With this commit every time a transaction is added, it is checked if any of the existing groups can be closed following that condition, if it can, the group is closed.